+++ system/info_linux.go
@@ -43,7 +43,7 @@ func GetSystemInfo() (*SystemInfo, error) {
 		Hostname:     byteSliceToString(info.Nodename),
 		OS:           byteSliceToString(info.Sysname),
 		Distribution: formatDistName(osInfo.Name),
-		Version:      osInfo.Version,
+		Version:      osInfo.VersionID,
 		Kernel:       byteSliceToString(info.Release),
 		Arch:         byteSliceToString(info.Machine),
 		ArchBits:     getCPUArchBits(),
@@ -52,7 +52,12 @@ func GetSystemInfo() (*SystemInfo, error) {
 
 // GetOSInfo returns info about OS
 func GetOSInfo() (*OSInfo, error) {
-	data, err := ioutil.ReadFile(osReleaseFile)
+	return ParseOSInfo(osReleaseFile)
+}
+
+// ParseOSInfo parses data in given os-release file
+func ParseOSInfo(file string) (*OSInfo, error) {
+	data, err := ioutil.ReadFile(file)
 
 	if err != nil {
 		return nil, err
@@ -68,34 +73,58 @@ func GetOSInfo() (*OSInfo, error) {
 		name := strutil.ReadField(line, 0, false, "=")
 		value := strings.Trim(strutil.ReadField(line, 1, false, "="), "\"")
 
-		switch name {
-		case "NAME":
-			info.Name = value
-		case "VERSION":
-			info.Version = value
-		case "ID":
-			info.ID = value
-		case "ID_LIKE":
-			info.IDLike = value
-		case "VERSION_ID":
-			info.VersionID = value
-		case "VERSION_CODENAME":
-			info.VersionCodename = value
-		case "PRETTY_NAME":
-			info.PrettyName = value
-		case "HOME_URL":
-			info.HomeURL = value
-		case "BUG_REPORT_URL":
-			info.BugReportURL = value
-		case "SUPPORT_URL":
-			info.SupportURL = value
-		}
+		applyOSInfo(info, name, value)
 	}
 
 	return info, nil
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
+
+// applyOSInfo applies record from os-release
+func applyOSInfo(info *OSInfo, name, value string) {
+	switch name {
+	case "NAME":
+		info.Name = value
+	case "VERSION":
+		info.Version = value
+	case "ID":
+		info.ID = value
+	case "ID_LIKE":
+		info.IDLike = value
+	case "PRETTY_NAME":
+		info.PrettyName = value
+	case "VERSION_ID":
+		info.VersionID = value
+	case "PLATFORM_ID":
+		info.PlatformID = value
+	case "VARIANT":
+		info.Variant = value
+	case "VARIANT_ID":
+		info.VariantID = value
+	case "CPE_NAME":
+		info.CPEName = value
+	case "VERSION_CODENAME":
+		info.VersionCodename = value
+	case "HOME_URL":
+		info.HomeURL = value
+	case "BUG_REPORT_URL":
+		info.BugReportURL = value
+	case "SUPPORT_URL":
+		info.SupportURL = value
+	case "DOCUMENTATION_URL":
+		info.DocumentationURL = value
+	case "LOGO":
+		info.Logo = value
+	}
+
+	switch {
+	case strings.HasSuffix(name, "SUPPORT_PRODUCT"):
+		info.SupportProduct = value
+	case strings.HasSuffix(name, "SUPPORT_PRODUCT_VERSION"):
+		info.SupportProductVersion = value
+	}
+}
 
 // formatDistName formats distribution name
 func formatDistName(name string) string {